refactor route manager into modules
