+++ index.html
@@ -205,6 +205,6 @@
     <script src="src/js/encryption-utils.js"></script>
     <script src="src/components/admin-panel-fix.js"></script>
     <script src="src/js/route-assignments.js"></script>
-    <script src="src/js/route-manager.js"></script>
+    <script type="module" src="src/js/route-manager.js"></script>
 </body>
 </html>
